Altered Right_ScaleIsRight and MoveElevatorAuto.
Right_ScaleIsRight now moves forward less initially and turns 45 degrees towards the scale, still need to test and implement in other Scale autos. MoveElevatorAuto now runs the elevator motors when the limit switch is not active, if active, it stops, but it continues to run the command until it times out.